+++ Model/Config/Backend/MinOrderTotal.php
@@ -1,0 +1,78 @@
+<?php
+
+/**
+ * HiPay Fullservice Magento
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Apache 2.0 Licence
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * @copyright Copyright (c) 2016 - HiPay
+ * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache 2.0 Licence
+ */
+
+namespace HiPay\FullserviceMagento\Model\Config\Backend;
+
+use Magento\Framework\App\Cache\TypeListInterface;
+use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\App\Config\Value;
+use Magento\Framework\Data\Collection\AbstractDb;
+use Magento\Framework\Exception\ValidatorException;
+use Magento\Framework\Model\Context;
+use Magento\Framework\Model\ResourceModel\AbstractResource;
+use Magento\Framework\Registry;
+
+class MinOrderTotal extends Value
+{
+    /**
+     * @var array
+     */
+    private $minimums;
+
+    /**
+     * @param Context               $context
+     * @param Registry              $registry
+     * @param ScopeConfigInterface  $config
+     * @param TypeListInterface     $cacheTypeList
+     * @param AbstractResource|null $resource
+     * @param AbstractDb|null       $resourceCollection
+     * @param array                 $data
+     * @param array                 $minimums
+     */
+    public function __construct(
+        Context $context,
+        Registry $registry,
+        ScopeConfigInterface $config,
+        TypeListInterface $cacheTypeList,
+        ?AbstractResource $resource = null,
+        ?AbstractDb $resourceCollection = null,
+        array $data = [],
+        array $minimums = []
+    ) {
+        $this->minimums = $minimums;
+        parent::__construct($context, $registry, $config, $cacheTypeList, $resource, $resourceCollection, $data);
+    }
+
+    /**
+     * @throws ValidatorException
+     * @return $this
+     */
+    public function beforeSave()
+    {
+        $min = (float) ($this->minimums[$this->getPath()] ?? 0.0);
+
+        if (
+            $min > 0 && $this->getValue() !== '' && $this->getValue() !== null
+            && (float) $this->getValue() < $min
+        ) {
+            throw new ValidatorException(
+                __('The minimum order total cannot be less than %1.', $min)
+            );
+        }
+
+        return parent::beforeSave();
+    }
+}