+++ view/frontend/web/js/view/payment/hipay-methods.js
@@ -71,6 +71,11 @@ define([
         'HiPay_FullserviceMagento/js/view/payment/method-renderer/hipay-bancomatpay-hosted-fields'
     },
     {
+      type: 'hipay_bizum_hosted_fields',
+      component:
+        'HiPay_FullserviceMagento/js/view/payment/method-renderer/hipay-bizum-hosted-fields'
+    },
+    {
       type: 'hipay_ideal_hosted_fields',
       component:
         'HiPay_FullserviceMagento/js/view/payment/method-renderer/hipay-ideal-hosted-fields'